backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
httpx==0.26.0
pydantic==2.5.3
pdfplumber==0.10.3
python-docx==1.1.0
aiofiles==23.2.1
